Move initialize simulation time function
